namada_ethereum_bridge::protocol::transactions::ethereum_events
Function apply_derived_tx

pub fn apply_derived_tx<D, H, Gov>(
    state: &mut WlState<D, H>,
    events: Vec<MultiSignedEthEvent>,
) -> Result<BatchedTxResult>
where
    D: 'static + DB + for<'iter> DBIter<'iter> + Sync,
    H: 'static + StorageHasher + Sync,
    Gov: Read<WlState<D, H>>,

Applies derived state changes to storage, based on Ethereum events which were newly seen by some consensus validator(s).

For events which have been seen by enough voting power (>= 2/3), extra state changes may take place, such as minting of wrapped ERC20s.

This function is deterministic based on some existing blockchain state and the passed events.