namada_ethereum_bridge::protocol::transactions
Module votes

Logic and data types relating to tallying validators’ votes for pieces of data stored in the ledger, where those pieces of data should only be acted on once they have received enough votes

Structs

Tally

Represents all the information needed to tally a piece of data that may be voted for over multiple epochs

Traits

EpochedVotingPowerExt

Extension methods for EpochedVotingPower instances.

Functions

calculate_new

Calculate a new Tally based on some validators’ fractional voting powers as specific block heights

dedupe

Deterministically constructs a Votes map from a set of validator addresses and the block heights they signed something at.

Type Aliases

EpochedVotingPower

The voting power behind a tally aggregated over multiple epochs.

Votes

The addresses of validators that voted for something, and the block heights at which they voted.