namada_ethereum_bridge::storage
Module eth_bridge_queries

Storage queries for ethereum bridge.

Structs

EthAssetMint

Number of tokens to mint after receiving a “transfer to Namada” Ethereum event.

EthBridgeQueriesHook

A handle to EthBridgeQueries.

Enums

EthBridgeEnabled

Enum indicating if the bridge was initialized at genesis or a later epoch.

EthBridgeStatus

An enum indicating if the Ethereum bridge is enabled.

SendValsetUpd

This enum is used as a parameter to EthBridgeQueriesHook::must_send_valset_upd.

Traits

EthBridgeQueries

Methods used to query blockchain Ethereum bridge related state.

Functions

check_bridge_status

Check if the bridge is disabled, enabled, or scheduled to be enabled at a specified Epoch.

is_bridge_active_at

Returns a boolean indicating whether the bridge is currently active at the specified Epoch.

is_bridge_comptime_enabled

Check if the Ethereum Bridge has been enabled at compile time.