namada_ethereum_bridge::storage
Module proof

Proofs over some arbitrary data.

Structs

EthereumProof

Ethereum proofs contain the secp256k1 signatures of validators over some data to be signed.

Functions

sort_sigs

Sort signatures based on voting powers in descending order. Puts a dummy signature in place of invalid or missing signatures.

Type Aliases

BridgePoolRootProof

Ethereum bridge pool root proof.