namada_ethereum_bridge::storage::proof
Function sort_sigs

pub fn sort_sigs(
    voting_powers: &VotingPowersMap,
    signatures: &HashMap<EthAddrBook, Signature>,
) -> Vec<Signature>

Sort signatures based on voting powers in descending order. Puts a dummy signature in place of invalid or missing signatures.