namada_ethereum_bridge::storage
Module vote_tallies

Functionality for accessing keys to do with tallying votes

Structs

BridgePoolRoot

A wrapper struct for managing keys related to tracking signatures over bridge pool roots and nonces.

Keys

Generator for the keys under which details of votes for some piece of data is stored

KeysSegments

Storage segments of Keys.

Constants

BRIDGE_POOL_ROOT_PREFIX_KEY_SEGMENT

Storage sub-key space reserved to keeping track of the voting power assigned to Ethereum bridge pool roots and nonces.

ETH_MSGS_PREFIX_KEY_SEGMENT

Storage sub-key space reserved to keeping track of the voting power assigned to Ethereum events.

VALSET_UPDS_PREFIX_KEY_SEGMENT

Storage sub-key space reserved to keeping track of the voting power assigned to validator set updates.

Functions

eth_event_keys

Get the Keys from the storage key. It returns None if the storage key isn’t for an Ethereum event.

eth_msgs_prefix

Get the key prefix corresponding to the storage location of EthereumEvents whose “seen” state is being tracked.

get_body_key_at_addr

get_seen_by_key_at_addr

get_seen_key_at_addr

get_voting_power_key_at_addr

get_voting_started_epoch_key_at_addr

is_body_key_at_addr

is_epoch_key

Return true if the storage key is a key to store the epoch

is_seen_by_key_at_addr

is_seen_key

Return true if the storage key is a key to store the seen

is_seen_key_at_addr

is_voting_power_key_at_addr

is_voting_started_epoch_key_at_addr

valset_upds_prefix

Get the key prefix corresponding to the storage location of validator set updates whose “seen” state is being tracked.